% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{make_anomalies}
\alias{make_anomalies}
\title{make anomaly features}
\usage{
make_anomalies(x, n = 0.01, strength = 1)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: 0.01}

\item{strength}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{sample}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}}
}
