% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{add_noise}
\alias{add_noise}
\title{add noise to a signal}
\usage{
add_noise(x, low = 0, high = 1)
}
\arguments{
\item{x}{a tibble from \code{\link{generate_signal}}}

\item{low}{lowest noise to add as a numeric, Default: 0}

\item{high}{highest noise to add as a numeric, Default: 1}
}
\value{
a tibble
}
\description{
add noise to a signal with set minimum and maximums
}
\examples{
\dontrun{
if(interactive()){
 generate_signal() \%>\% add_noise()
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
