% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{add_seasonality}
\alias{add_seasonality}
\title{add seasonality to a signal}
\usage{
add_seasonality(x, seasonality_strength = 5)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{seasonality_strength}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
