% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{add_trend}
\alias{add_trend}
\title{add trend to a signal}
\usage{
add_trend(x, strength = 1, start = NA)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{trend_strength}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
