% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{make_weekends}
\alias{make_weekends}
\title{make weekend features}
\usage{
make_weekends(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
\code{\link[lubridate]{day}}
\code{\link[tsibble]{as_tsibble}},\code{\link[tsibble]{slide}}
}
