% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_wide}
\alias{format_wide}
\title{format tidy data to industry wide format}
\usage{
format_wide(x)
}
\arguments{
\item{x}{a tibble from \code{\link{generate_signal}}}
}
\value{
a tibble
}
\description{
format tidy data to industry wide format with columns as half hours and rows as days
}
\examples{
\dontrun{
if(interactive()){
 generate_signal() \%>\% format_wide()
 }
}
}
\seealso{
\code{\link[lubridate]{date}}
}
