% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{make_anomalies}
\alias{make_anomalies}
\title{make anomaly features}
\usage{
make_anomalies(x, n = 0.01, strength = 1)
}
\arguments{
\item{x}{a tibble from \code{\link{generate_signal}}}

\item{n}{the percentage of the data to insert anomalies into as a numeric <=
1, Default: 0.01}

\item{strength}{the strength of the anomalies in the signal as a numeric,
Default: 1}
}
\value{
a tibble
}
\description{
make anomaly features of varying number that randomly increase
the signal by a specific volume
}
\examples{
\dontrun{
if(interactive()){
 generate_signal() \%>\% make_anomalies()
 }
}
}
\seealso{
\code{\link[dplyr]{sample}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}}
}
