% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{make_trend}
\alias{make_trend}
\title{make trend features}
\usage{
make_trend(x, strength = 10, start = NA)
}
\arguments{
\item{x}{a tibble from \code{\link{generate_signal}}}

\item{strength}{the strength of the trend in the signal as a numeric,
Default: 1}

\item{start}{interval to start on as an integer, Default: NA}
}
\value{
a tibble
}
\description{
make trend features that start at a specific interval that
increase or decrease the signal linearly by a specific strength
}
\examples{
\dontrun{
if(interactive()){
 generate_signal() \%>\% make_trend()
 }
}
}
