% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{make_weekends}
\alias{make_weekends}
\title{make weekend features}
\usage{
make_weekends(x)
}
\arguments{
\item{x}{a tibble from \code{\link{generate_signal}}}
}
\value{
a tssibble
}
\description{
make weekend features for saturdays and sundays that are capped
in value to the local average
}
\examples{
\dontrun{
if(interactive()){
 generate_signal() \%>\% make_weekends()
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}} \code{\link[lubridate]{day}}
\code{\link[tsibble]{as_tsibble}},\code{\link[tsibble]{slide}}
}
