% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{generate_signal}
\alias{generate_signal}
\title{generate a diurnal signal of half hours}
\usage{
generate_signal(days = 7 * 4, start = "2020-01-01")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
}
