% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{shift_signal}
\alias{shift_signal}
\title{shift a signal by a margin}
\usage{
shift_signal(x, centre_point)
}
\arguments{
\item{x}{a tibble from \code{\link{generate_signal}}}

\item{centre_point}{the amount to shift up by as a numeric}
}
\value{
a tibble
}
\description{
shift signal by a margin up or down to represent magnitude of
the value
}
\examples{
\dontrun{
if(interactive()){
 generate_signal(5) \%>\% shift_signal(9)
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
