% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{add_seasonality}
\alias{add_seasonality}
\title{add seasonality to a signal}
\usage{
add_seasonality(x, strength = 5, invert = FALSE)
}
\arguments{
\item{x}{a tibble from \code{\link{generate_signal}}}

\item{strength}{the strength of the change in a seasonal cycle as a numeric,
Default: 5}

\item{invert}{if the signal should be higher in summer than it is in winter
as a booleen, Default: FALSE}
}
\value{
a tibble
}
\description{
add seasonality to a signal of a specific strength, that can be
high in summer or high in winter, that is exactly a year long
}
\examples{
\dontrun{
if(interactive()){
 generate_signal() \%>\%  add_seasonality()
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
