% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{generate_signal}
\alias{generate_signal}
\title{generate a diurnal signal of half hours}
\usage{
generate_signal(days, start = "2020", amplitude = NA)
}
\arguments{
\item{days}{PARAM_DESCRIPTION, Default: 7 * 4}

\item{start}{PARAM_DESCRIPTION, Default: "2020"}

\item{amplitude}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
}
