% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{generate_signal}
\alias{generate_signal}
\title{generate a diurnal signal of half hours}
\usage{
generate_signal(days, start = "2020", amplitude = NA)
}
\arguments{
\item{days}{number of days to generate signal for as an integer, Default: 7 *
4}

\item{start}{the year to start from as a string of `yyyy``, Default: "2020"}

\item{amplitude}{the height of the signal as a numeric, Default: NA}
}
\value{
a tibble
}
\description{
generate a diurnal signal of half hours over a range of days
starting from the first of the selected year, with varying amplitude
}
\examples{
\dontrun{
if(interactive()){
 generate_signal()

 generate_signal(365, start = "2019", amplitude = 12)
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
}
