% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{pure_signal}
\alias{pure_signal}
\title{generate a diurnal signal of half hours}
\usage{
pure_signal(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
}
