% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signal.R
\name{shift_signal}
\alias{shift_signal}
\title{shift a signal by a margin}
\usage{
shift_signal(x, centre_point)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{centre_point}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
